requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Platform Integrations
requests==2.31.0
playwright==1.41.0
selenium==4.17.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Google Sheets
gspread==5.12.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# OAuth & Security
authlib==1.3.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
aiofiles==23.2.1
pillow==10.2.0

# Task Scheduling
apscheduler==3.10.4
celery==5.3.6
redis==5.0.1

# Development
pytest==7.4.4
black==24.1.1
flake8==7.0.0
